+++ cli/tests/runner.rs
@@ -57,6 +57,7 @@ mod test_identical_commits;
 mod test_immutable_commits;
 mod test_interdiff_command;
 mod test_log_command;
+mod test_mailmap;
 mod test_metaedit_command;
 mod test_new_command;
 mod test_next_prev_commands;