+++ lib/src/lib.rs
@@ -70,6 +70,7 @@ pub mod index;
 pub mod iter_util;
 pub mod local_working_copy;
 pub mod lock;
+pub mod mailmap;
 pub mod matchers;
 pub mod merge;
 pub mod merged_tree;