+++ lib/tests/test_revset.rs
@@ -32,6 +32,8 @@ use jj_lib::git;
 use jj_lib::graph::GraphEdge;
 use jj_lib::graph::reverse_graph;
 use jj_lib::id_prefix::IdPrefixContext;
+use jj_lib::mailmap::Mailmap;
+use jj_lib::mailmap::read_current_mailmap;
 use jj_lib::merge::Merge;
 use jj_lib::merged_tree::MergedTree;
 use jj_lib::object_id::ObjectId as _;
@@ -102,6 +104,7 @@ fn resolve_symbol(repo: &dyn Repo, symbol: &str) -> Result<Vec<CommitId>, Revset
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: None,
+        mailmap: Arc::new(Mailmap::empty()),
     };
     let expression = parse(&mut RevsetDiagnostics::new(), symbol, &context).unwrap();
     assert_matches!(*expression, RevsetExpression::CommitRef(_));
@@ -236,6 +239,7 @@ fn test_resolve_symbol_commit_id() {
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: None,
+        mailmap: Arc::new(Mailmap::empty()),
     };
     assert_matches!(
         parse(&mut RevsetDiagnostics::new(), "present(01)", &context).unwrap()
@@ -1083,12 +1087,17 @@ fn test_resolve_symbol_git_refs() {
     );
 }
 
-fn resolve_commit_ids(repo: &dyn Repo, revset_str: &str) -> Vec<CommitId> {
-    try_resolve_commit_ids(repo, revset_str).unwrap()
+fn resolve_commit_ids_with_mailmap(
+    repo: &dyn Repo,
+    mailmap_source: &str,
+    revset_str: &str,
+) -> Vec<CommitId> {
+    try_resolve_commit_ids_with_mailmap(repo, mailmap_source, revset_str).unwrap()
 }
 
 fn try_resolve_expression(
     repo: &dyn Repo,
+    mailmap_source: &str,
     revset_str: &str,
 ) -> Result<Arc<ResolvedRevsetExpression>, RevsetResolutionError> {
     let settings = testutils::user_settings();
@@ -1101,17 +1110,19 @@ fn try_resolve_expression(
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: None,
+        mailmap: Arc::new(Mailmap::from_bytes(mailmap_source.as_bytes())),
     };
     let expression = parse(&mut RevsetDiagnostics::new(), revset_str, &context).unwrap();
     let symbol_resolver = default_symbol_resolver(repo);
     expression.resolve_user_expression(repo, &symbol_resolver)
 }
 
-fn try_resolve_commit_ids(
+fn try_resolve_commit_ids_with_mailmap(
     repo: &dyn Repo,
+    mailmap_source: &str,
     revset_str: &str,
 ) -> Result<Vec<CommitId>, RevsetResolutionError> {
-    Ok(try_resolve_expression(repo, revset_str)?
+    Ok(try_resolve_expression(repo, mailmap_source, revset_str)?
         .evaluate(repo)
         .unwrap()
         .iter()
@@ -1119,11 +1130,22 @@ fn try_resolve_commit_ids(
         .collect())
 }
 
+fn resolve_commit_ids(repo: &dyn Repo, revset_str: &str) -> Vec<CommitId> {
+    resolve_commit_ids_with_mailmap(repo, "", revset_str)
+}
+
+fn try_resolve_commit_ids(
+    repo: &dyn Repo,
+    revset_str: &str,
+) -> Result<Vec<CommitId>, RevsetResolutionError> {
+    try_resolve_commit_ids_with_mailmap(repo, "", revset_str)
+}
+
 fn try_evaluate_expression<'index>(
     repo: &'index dyn Repo,
     revset_str: &str,
 ) -> Result<Box<dyn Revset + 'index>, RevsetEvaluationError> {
-    try_resolve_expression(repo, revset_str)
+    try_resolve_expression(repo, "", revset_str)
         .unwrap()
         .evaluate(repo)
 }
@@ -1152,6 +1174,11 @@ fn resolve_commit_ids_in_workspace(
         use_glob_by_default: true,
         extensions: &RevsetExtensions::default(),
         workspace: Some(workspace_ctx),
+        mailmap: Arc::new(
+            read_current_mailmap(repo, workspace_ctx.workspace_name)
+                .block_on()
+                .unwrap(),
+        ),
     };
     let expression = parse(&mut RevsetDiagnostics::new(), revset_str, &context).unwrap();
     let symbol_resolver = default_symbol_resolver(repo);
@@ -3499,6 +3526,20 @@ fn test_evaluate_expression_author() {
         ),
         vec![commit3.id().clone(), commit1.id().clone()]
     );
+    // Signatures are treated as their mailmapped forms
+    let mailmap = "nameone <email1>";
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(mut_repo, mailmap, "author(\"nameone\")"),
+        vec![commit1.id().clone()]
+    );
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(mut_repo, mailmap, "author(\"name1\")"),
+        vec![]
+    );
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(mut_repo, mailmap, "author_raw(\"name1\")"),
+        vec![commit1.id().clone()]
+    );
 }
 
 fn parse_timestamp(s: &str) -> Timestamp {
@@ -3702,6 +3743,16 @@ fn test_evaluate_expression_mine() {
             commit1.id().clone()
         ]
     );
+    // Signatures are treated as their mailmapped forms
+    let user_email = settings.user_email();
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(
+            mut_repo,
+            format!("<{user_email}> <email1>\n<nobody> name2 <{user_email}>").as_ref(),
+            "mine()"
+        ),
+        vec![commit3.id().clone(), commit1.id().clone()]
+    );
 }
 
 #[test]
@@ -3841,6 +3892,20 @@ fn test_evaluate_expression_committer() {
     assert_eq!(
         resolve_commit_ids(mut_repo, "visible_heads() & committer(*name2*)"),
         vec![]
+    );
+    // Signatures are treated as their mailmapped forms
+    let mailmap = "nameone <email1>";
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(mut_repo, mailmap, "committer(\"nameone\")"),
+        vec![commit1.id().clone()]
+    );
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(mut_repo, mailmap, "committer(\"name1\")"),
+        vec![]
+    );
+    assert_eq!(
+        resolve_commit_ids_with_mailmap(mut_repo, mailmap, "committer_raw(\"name1\")"),
+        vec![commit1.id().clone()]
     );
 }